use intptr_t in align4, removing a truncation warning #29

+++ zzip/zip.c
@@ -388,8 +388,8 @@ _zzip_inline static char *
 __zzip_aligned4(char *p)
 {
 #define aligned4   __zzip_aligned4
-    p += ((long) p) & 1;        /* warnings about truncation of a "pointer" */
-    p += ((long) p) & 2;        /* to a "long int" may be safely ignored :) */
+    p += ((intptr_t) p) & 1;
+    p += ((intptr_t) p) & 2;
     return p;
 }
